+++ src/main/java/net/rpgdifficulty/mixin/compat/ChangeAttributeMixin.java
@@ -45,7 +45,7 @@ public class ChangeAttributeMixin {
             int playerCount = 0;
             int totalPlayerLevel = 0;
             for (PlayerEntity playerEntity : world.getPlayers()) {
-                if (!EntityPredicates.EXCEPT_CREATIVE_OR_SPECTATOR.test(playerEntity)) {
+                if (EntityPredicates.EXCEPT_CREATIVE_OR_SPECTATOR.test(playerEntity)) {
                     continue;
                 }
                 if (playerEntity.getWorld().getDimension().equals(mobEntity.getWorld().getDimension())